+++ cmp/dock/DockViewModel.js
@@ -5,10 +5,9 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 import {HoistModel, managed, XH} from '@xh/hoist/core';
+import {RefreshModeContextModel} from '@xh/hoist/core/refresh';
 import {bindable, observable, action} from '@xh/hoist/mobx';
 import {throwIf} from '@xh/hoist/utils/js';
-
-import {DockViewRefreshContextModel} from './impl/DockViewRefreshContextModel';
 
 /**
  * Model for a DockView within a DockContainer. Specifies the actual content (child component)
@@ -94,7 +93,7 @@ export class DockViewModel {
         this._renderMode = renderMode;
         this._refreshMode = refreshMode;
 
-        this.refreshContextModel = new DockViewRefreshContextModel(this);
+        this.refreshContextModel = new RefreshModeContextModel(this);
     }
 
     //-----------------------