+++ desktop/cmp/panel/PanelModel.js
@@ -5,7 +5,8 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 
-import {XH, HoistModel, RenderMode} from '@xh/hoist/core';
+import {XH, HoistModel, managed, RenderMode, RefreshMode} from '@xh/hoist/core';
+import {RefreshModeContextModel} from '@xh/hoist/core/refresh';
 import {observable, action} from '@xh/hoist/mobx';
 import {withDefault} from '@xh/hoist/utils/js';
 import {start} from '@xh/hoist/promise';
@@ -29,11 +30,14 @@ export class PanelModel {
     maxSize;
     defaultCollapsed;
     side;
-    collapsedRenderMode;
+    renderMode;
+    refreshMode;
     prefName;
     showSplitter;
     showSplitterCollapseButton;
     showHeaderCollapseButton;
+
+    @managed refreshContextModel;
 
     //---------------------
     // Observable State
@@ -48,6 +52,13 @@ export class PanelModel {
     @observable isResizing = false;
 
     /**
+     * RenderMode and RefreshMode consider this panel 'active' when it is not collapsed
+     */
+    get isActive() {
+        return !this.collapsed;
+    }
+
+    /**
      * @param {Object} config
      * @param {boolean} [config.resizable] - Can panel be resized?
      * @param {boolean} [config.resizeWhileDragging] - Redraw panel as resize happens?
@@ -58,7 +69,8 @@ export class PanelModel {
      * @param {boolean} [config.defaultCollapsed] - Default collapsed state.
      * @param {string} config.side - Side towards which the panel collapses or shrinks. This relates
      *      to the position within a parent vbox or hbox in which the panel should be placed.
-     * @param {RenderMode} [config.collapsedRenderMode] - How should collapsed content be rendered?
+     * @param {RenderMode} [config.renderMode] - How should collapsed content be rendered?
+     * @param {RefreshMode} [config.refreshMode] - How should collapsed content be refreshed?
      * @param {?string} [config.prefName] - preference name to store sizing and collapsed state.
      * @param {boolean} [config.showSplitter] - Should a splitter be rendered at the panel edge?
      * @param {boolean} [config.showSplitterCollapseButton] - Should the collapse button be visible
@@ -75,7 +87,8 @@ export class PanelModel {
         maxSize = null,
         defaultCollapsed = false,
         side,
-        collapsedRenderMode = RenderMode.LAZY,
+        renderMode = RenderMode.LAZY,
+        refreshMode = RefreshMode.ON_SHOW_LAZY,
         prefName = null,
         showSplitter = resizable || collapsible,
         showSplitterCollapseButton = showSplitter && collapsible,
@@ -102,10 +115,13 @@ export class PanelModel {
         this.maxSize = maxSize;
         this.defaultCollapsed = defaultCollapsed;
         this.side = side;
-        this.collapsedRenderMode = collapsedRenderMode;
+        this.renderMode = renderMode;
+        this.refreshMode = refreshMode;
         this.showSplitter = showSplitter;
         this.showSplitterCollapseButton = showSplitterCollapseButton;
         this.showHeaderCollapseButton = showHeaderCollapseButton;
+
+        this.refreshContextModel = new RefreshModeContextModel(this);
 
         if (prefName && !XH.prefService.hasKey(prefName)) {
             console.warn(`Unknown preference for storing state of Panel '${prefName}'`);